Pruebas en SaveFile custom values

+++ src/domain/use-cases/save-file.use.case.test.ts
@@ -4,10 +4,18 @@ import { SaveFile } from "./save-file.use.case"
 
 describe('SaveFileUseCase', () => { 
 
+    const customOptions = {
+        fileContent: 'custom-content',
+        fileDestination: 'custom-outputs/file-destination',
+        fileName:'custom-table-name'
+    }
+    const customFilePath =`${customOptions.fileDestination}/${customOptions.fileName}.txt`;
     
-
     afterEach( () => {
-        fs.rmSync('outputs', {recursive: true})
+        const outputFolderExists = fs.existsSync('outputs')
+        const outputCustomFolderExists = fs.existsSync(customOptions.fileDestination)
+        if(outputFolderExists) fs.rmSync('outputs', {recursive: true})
+        if(outputCustomFolderExists) fs.rmSync('custom-outputs', {recursive: true})
     })
 
 
@@ -29,4 +37,21 @@ describe('SaveFileUseCase', () => {
     
 
      })
+
+     test('should save file with custom values', () => { 
+      
+        const saveFile = new SaveFile();
+        const result = saveFile.execute(customOptions)
+
+        const fileExists =fs.existsSync(customFilePath)
+        const fileContent =fs.readFileSync(customFilePath, 'utf-8')
+        
+        expect(result).toBe(true)
+        expect(fileExists).toBe(true)
+        expect(fileContent).toBe(customOptions.fileContent)
+        
+      })
+
+
+      
  })
